requirements.txt
fastapi==0.115.5
uvicorn==0.30.6
mlflow==2.10.2
pandas==2.1.4
numpy==1.26.4
scikit-learn==1.3.2
scipy==1.11.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1
